impl - delete item permissions

+++ backend/src/resolvers/Mutation.js
@@ -46,11 +46,18 @@ const mutations = {
         );
     },
     async deleteItem(parent, args, ctx, info){
+        throw new Error("You're not allowed");
         const where = {id: args.id};
         // find item
-        const item = await ctx.db.query.item({where}, `{id title }`);
+        const item = await ctx.db.query.item({where}, `{ id title user { id }}`);
         // check if owned / permissons
-        //TODO
+        const ownsItem = item.user.id = ctx.request.userId;
+        const hasPermissions = ctx.request.user.permissions.some(
+            permission => ['ADMIN', 'ITEMDELETE'].includes(permission)
+        )
+        if(!ownsItem || !hasPermissions){
+            throw new Error("You dont have permissions")
+        }
         // delete item
         return ctx.db.mutation.deleteItem({ where }, info)
     },
